v1.1.0 refactring for relocalization. (works without optimization)

+++ include/se2lam/MapStorage.h
@@ -31,7 +31,7 @@ public:
 
     struct CovisibleRelationship
     {
-        unsigned long thisKFid, thatKFid;
+        int thisKFid, thatKFid;
         int covisbleCount;
     };
